feat: clean up old records auth-accp
